Move isFatalSessionError to shared/errors
