+++ apps/twig/src/shared/errors.ts
@@ -19,3 +19,29 @@ export function isAuthError(error: unknown): boolean {
   if (!message) return false;
   return AUTH_ERROR_PATTERNS.some((pattern) => message.includes(pattern));
 }
+
+const FATAL_SESSION_ERROR_PATTERNS = [
+  "Internal error",
+  "process exited",
+  "Session did not end",
+  "not ready for writing",
+  "Session not found",
+] as const;
+
+function includesAny(
+  value: string | undefined,
+  patterns: readonly string[],
+): boolean {
+  if (!value) return false;
+  return patterns.some((pattern) => value.includes(pattern));
+}
+
+export function isFatalSessionError(
+  errorMessage: string,
+  errorDetails?: string,
+): boolean {
+  return (
+    includesAny(errorMessage, FATAL_SESSION_ERROR_PATTERNS) ||
+    includesAny(errorDetails, FATAL_SESSION_ERROR_PATTERNS)
+  );
+}
